+++ baybe/recommenders/pure/nonpredictive/sampling.py
@@ -1,9 +1,11 @@
 """Recommenders based on sampling."""
 
-from typing import ClassVar
+from typing import ClassVar, Literal
 
 import numpy as np
 import pandas as pd
+from attrs import define, field
+from attrs.validators import instance_of
 from sklearn.preprocessing import StandardScaler
 from typing_extensions import override
 
@@ -52,12 +54,18 @@ class RandomRecommender(NonPredictiveRecommender):
         return to_string(self.__class__.__name__, *fields)
 
 
+@define
 class FPSRecommender(NonPredictiveRecommender):
     """An initial recommender that selects candidates via Farthest Point Sampling."""
 
     # Class variables
     compatibility: ClassVar[SearchSpaceType] = SearchSpaceType.DISCRETE
     # See base class.
+
+    initialization: Literal["farthest", "random"] = field(
+        default="farthest", validator=instance_of(str)
+    )
+    random_tie_break: bool = field(default=True, validator=instance_of(bool))
 
     @override
     def _recommend_discrete(
@@ -74,7 +82,12 @@ class FPSRecommender(NonPredictiveRecommender):
         # Scale and sample
         candidates_comp = subspace_discrete.transform(candidates_exp)
         candidates_scaled = np.ascontiguousarray(scaler.transform(candidates_comp))
-        ilocs = farthest_point_sampling(candidates_scaled, batch_size)
+        ilocs = farthest_point_sampling(
+            candidates_scaled,
+            batch_size,
+            initialization=self.initialization,
+            random_tie_break=self.random_tie_break,
+        )
         return candidates_comp.index[ilocs]
 
     @override