+++ CHANGELOG.md
@@ -70,6 +70,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `ngboost` is now an optional dependency
 - `setuptools-scm` is now an optional dependency, used for improved version inference
 - `create_example_plots`, `to_string` and `indent` have been relocated within utils
+- `random_tie_break` flag added to `farthest_point_sampling`, that toggles between 
+  random or deterministic sampling for equidistant cases
 
 ### Fixed
 - Incorrect optimization direction with `PSTD` with a single minimization target